Rename to Analytics.h
